Connection Pool + Transaction implemented
Dependency added for connection Pool
