--- Case 1 - ORM Spring Java/case1/src/main/java/com/example/case1/Repositories/sd_ownersRepository.java
@@ -1,13 +1,0 @@
-package com.example.case1.Repositories;
-
-import java.util.List;
-import com.example.case1.Entities.sd_owners;
-
-import org.springframework.data.repository.CrudRepository;
-import org.springframework.stereotype.Repository;
-
-@Repository
-public interface sd_ownersRepository extends CrudRepository<sd_owners, Integer> {
-    List<sd_owners> findBylastname(String pLastName);
-    sd_owners findByownerid(long pId);
-}
