+++ Case 1 - ORM Spring Java/case1/src/main/resources/hibernate.cfg.xml
@@ -11,5 +11,14 @@
         <property name="hibernate.connection.password">Guayaboscr123</property>
         <property name="hibernate.current_session_context_class">thread</property>
         <property name="hibernate.show_sql">true</property>
+
+        <property name="hibernate.c3p0.min_size">5</property>
+        <property name="hibernate.c3p0.max_size">20</property>
+        <property name="hibernate.c3p0.timeout">300</property>
+        <property name="hibernate.c3p0.max_statements">50</property>
+        <property name="hibernate.c3p0.idle_test_period">3000</property>
+
+        <mapping class="com.example.case1.Entities.Owner"/>
+        <mapping class="com.example.case1.Entities.Problem"/>
     </session-factory>
 </hibernate-configuration>
